moved UpdateCompanyBusinessTypeModal
